refactor(CSI-324): break down API client into multiple files by functional area
